+++ apps/mobile/src/components/posts/card.tsx
@@ -176,6 +176,7 @@ export function PostCard({ expanded, post, style, viewing }: Props) {
           onPress={onPress}
           post={post}
           side={mediaOnRight ? 'right' : 'left'}
+          viewing={viewing}
         />
       </PostGestures>
     )
@@ -249,6 +250,7 @@ export function PostCard({ expanded, post, style, viewing }: Props) {
           recyclingKey={post.id}
           spoiler={post.spoiler}
           style={body ? styles.expanded : null}
+          viewing={viewing}
         />
       ) : null}